+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release Notes for Read Time
 
+## 1.0.1 2025-01-24
+
+### Changed
+
+- Moved field processing logic into a service
+- Save readtime value as an integer instead of a string
+- Limit processing to text-based fields
+- Updated README
+
+### Added
+
+- Plugin icon
+- Formatting twig extension for integer value
+
 ## 1.0.0 2025-01-23
 
 - Initial commit